requirements.txt
Flask
ytmusicapi
python-dotenv
yt-dlp